+++ shardingsphere-infra/shardingsphere-infra-mode/shardingsphere-infra-mode-type/shardingsphere-infra-standalone-mode/shardingsphere-infra-standalone-mode-core/src/main/java/org/apache/shardingsphere/infra/mode/manager/standalone/StandaloneContextManagerBuilder.java
@@ -28,8 +28,8 @@ import org.apache.shardingsphere.infra.context.metadata.MetaDataContextsBuilder;
 import org.apache.shardingsphere.infra.metadata.resource.ShardingSphereResource;
 import org.apache.shardingsphere.infra.mode.config.ModeConfiguration;
 import org.apache.shardingsphere.infra.mode.config.PersistRepositoryConfiguration;
-import org.apache.shardingsphere.infra.mode.manager.standalone.repository.StandalonePersistRepository;
-import org.apache.shardingsphere.infra.mode.manager.standalone.repository.StandalonePersistRepositoryConfiguration;
+import org.apache.shardingsphere.infra.mode.repository.standalone.StandalonePersistRepository;
+import org.apache.shardingsphere.infra.mode.repository.standalone.StandalonePersistRepositoryConfiguration;
 import org.apache.shardingsphere.infra.persist.PersistService;
 import org.apache.shardingsphere.infra.spi.ShardingSphereServiceLoader;
 import org.apache.shardingsphere.infra.spi.typed.TypedSPIRegistry;
