+++ shardingsphere-infra/shardingsphere-infra-mode/shardingsphere-infra-mode-type/shardingsphere-infra-standalone-mode/shardingsphere-infra-standalone-mode-repository/shardingsphere-infra-standalone-mode-repository-api/src/main/java/org/apache/shardingsphere/infra/mode/repository/standalone/StandalonePersistRepositoryConfiguration.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.mode.manager.standalone.repository;
+package org.apache.shardingsphere.infra.mode.repository.standalone;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
