+++ shardingsphere-infra/shardingsphere-infra-mode/shardingsphere-infra-mode-type/shardingsphere-infra-standalone-mode/shardingsphere-infra-standalone-mode-repository/shardingsphere-infra-standalone-mode-repository-provider/shardingsphere-infra-standalone-mode-local-repository/src/main/java/org/apache/shardingsphere/infra/mode/repository/standalone/local/LocalRepository.java
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.mode.manager.standalone.repository.local;
+package org.apache.shardingsphere.infra.mode.repository.standalone.local;
 
 import com.google.common.base.Joiner;
 import com.google.common.base.Strings;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.shardingsphere.infra.mode.manager.standalone.repository.StandalonePersistRepository;
+import org.apache.shardingsphere.infra.mode.repository.standalone.StandalonePersistRepository;
 
 import java.io.BufferedWriter;
 import java.io.File;
