+++ shardingsphere-infra/shardingsphere-infra-mode/shardingsphere-infra-mode-type/shardingsphere-infra-standalone-mode/shardingsphere-infra-standalone-mode-repository/shardingsphere-infra-standalone-mode-repository-provider/shardingsphere-infra-standalone-mode-local-repository/src/main/java/org/apache/shardingsphere/infra/mode/repository/standalone/local/LocalRepositoryPropertyKey.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.mode.manager.standalone.repository.local;
+package org.apache.shardingsphere.infra.mode.repository.standalone.local;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
